requirements.txt
alembic==0.9.6
asn1crypto==0.23.0
attrs==17.3.0
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
click==6.7
colorama==0.3.9
cryptography==2.1.4
ecdsa==0.13
Flask==0.12.2
Flask-Login==0.4.0
git+https://github.com/jfinkels/flask-restless/
Flask-SQLAlchemy==2.3.2
future==0.16.0
gunicorn==19.7.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
marshmallow==2.14.0
marshmallow-jsonapi==0.16.0
mimerender==0.6.0
pluggy==0.6.0
py==1.5.2
pycparser==2.18
pycrypto==2.6.1
PyMySQL==0.7.11
pytest==3.3.0
pytest-flask==0.10.0
python-dateutil==2.6.1
python-editor==1.0.3
python-jose==1.4.0
python-mimeparse==1.6.0
requests==2.18.4
six==1.11.0
SQLAlchemy==1.1.15
urllib3==1.22
Werkzeug==0.12.2
WTForms==2.1
